refactor(ocr): remove MCP server, run inference directly in extension process\n\n- Merge all ONNX inference logic (preprocessing, DB post-process, CTC\n  decoding) from mcp-server.ts into index.ts\n- Delete src/mcp-server.ts — extension runs fully in-process, no MCP\n  child process lifecycle to manage\n- Remove @modelcontextprotocol/sdk dependency\n- Simplify build.mjs to a single esbuild target\n- Update i18n: remove MCP-specific messages, add notConfigured error\n- Rewrite READMEs to reflect the new architecture
