+++ extensions/ocr/i18n/en-US.json
@@ -24,17 +24,15 @@
   "form.setup.language.options.ch": "Chinese",
   "form.setup.language.options.ch+en+ja": "Chinese + English + Japanese",
   "toast.saved.title": "OCR configuration saved",
-  "toast.saved.description": "PP-OCRv6 models are ready. MCP Client will start the OCR service.",
+  "toast.saved.description": "PP-OCRv6 models are ready and loaded into memory.",
   "status.noModels": "**No models cached.** Run `setup_ocr` to download models.",
-  "status.mcpUnavailable": "**MCP Client:** not available (enable the MCP Client extension)",
   "status.cachedModels": "**Cached models:**",
   "status.charsCount": "Character dictionary: {count} entries",
-  "status.mcpServer": "**OCR MCP Server:** {status}",
-  "status.availableTools": "**Available tools:** {tools}",
   "error.downloadFailed": "Failed to download OCR models: {message}",
   "error.nativeDepsFailed": "Failed to install native dependencies (onnxruntime-node / sharp): {message}",
-  "error.configSaveFailed": "Failed to save OCR MCP config: {message}",
+  "error.configSaveFailed": "Failed to save OCR config: {message}",
   "cancelled": "OCR setup was cancelled ({reason})",
+  "error.notConfigured": "PP-OCRv6 is not configured yet. Run `setup_ocr` first.",
   "form.ocrImage.title": "Select Image",
   "form.ocrImage.description": "Enter or drag an image file path — PP-OCRv6 will extract text from it.",
   "form.ocrImage.submit": "Start OCR",
@@ -42,7 +40,6 @@
   "form.ocrImage.imagePath.placeholder": "/path/to/image.png or drag file here",
   "error.noImagePath": "No image path provided.",
   "error.fileNotFound": "File not found: {path}",
-  "error.mcpUnavailable": "MCP Client is not available. Please enable the MCP Client extension first.",
   "error.ocrFailed": "OCR recognition failed: {message}",
   "error.ocrNoResult": "OCR returned no result.",
   "settings.language.label": "Language"
