+++ extensions/ocr/i18n/zh-CN.json
@@ -24,17 +24,15 @@
   "form.setup.language.options.ch": "中文",
   "form.setup.language.options.ch+en+ja": "中文 + 英文 + 日文",
   "toast.saved.title": "OCR 配置已保存",
-  "toast.saved.description": "PP-OCRv6 模型已就绪，MCP Client 将启动 OCR 服务。",
+  "toast.saved.description": "PP-OCRv6 模型已就绪，已加载到内存。",
   "status.noModels": "**没有缓存的模型。** 运行 setup_ocr 下载模型。",
-  "status.mcpUnavailable": "**MCP Client:** 不可用（请启用 MCP Client 扩展）",
   "status.cachedModels": "**已缓存的模型:**",
   "status.charsCount": "字符字典: {count} 个条目",
-  "status.mcpServer": "**OCR MCP 服务:** {status}",
-  "status.availableTools": "**可用工具:** {tools}",
   "error.downloadFailed": "下载 OCR 模型失败: {message}",
   "error.nativeDepsFailed": "安装原生依赖失败（onnxruntime-node / sharp）: {message}",
-  "error.configSaveFailed": "保存 OCR MCP 配置失败: {message}",
+  "error.configSaveFailed": "保存 OCR 配置失败: {message}",
   "cancelled": "OCR 设置已取消 ({reason})",
+  "error.notConfigured": "PP-OCRv6 尚未配置，请先运行 setup_ocr。",
   "form.ocrImage.title": "选择图片",
   "form.ocrImage.description": "输入或拖放图片文件路径，PP-OCRv6 将提取其中的文字。",
   "form.ocrImage.submit": "开始识别",
@@ -42,7 +40,6 @@
   "form.ocrImage.imagePath.placeholder": "/path/to/image.png 或拖放文件到此处",
   "error.noImagePath": "未提供图片路径。",
   "error.fileNotFound": "文件不存在: {path}",
-  "error.mcpUnavailable": "MCP Client 未启用。请先启用 MCP Client 扩展。",
   "error.ocrFailed": "OCR 识别失败: {message}",
   "error.ocrNoResult": "OCR 未返回结果。",
   "settings.language.label": "识别语言"
