+++ extensions/ocr/package.json
@@ -47,23 +47,8 @@
     ],
     "privacyPolicyUrl": "https://finchwork.app/privacy",
     "termsOfServiceUrl": "https://finchwork.app/terms",
-    "requires": {
-      "capabilities": [
-        "mcp.client"
-      ]
-    },
     "contributes": {
-      "tools": true,
-      "mcpServers": [
-        {
-          "name": "ocr",
-          "command": "node",
-          "args": [
-            "dist/mcp-server.js"
-          ],
-          "description": "PP-OCRv6 MCP server — provides ocr_image, ocr_languages, ocr_status tools."
-        }
-      ]
+      "tools": true
     },
     "settings": {
       "fields": [
@@ -101,7 +86,6 @@
     }
   },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.29.0",
     "onnxruntime-node": "^1.22.0",
     "sharp": "^0.34.0"
   },
